Fixed UI loading issue, allowed users to be passed between pages
The Program boots into LoginPage -> MainPage - > AddImage without any errors so far. You are now able to obtain access the list of users, that remain in memory throughout the usage of the app. Allows access to user images and information. Added in functionality to get the currently logged in user.