Selectable images, Cancel Works as intended, Can Log out,
I have replaced the Vbox with a ListView this gives the ability to scroll aswell as the option to select a specific image. Instead of removing the Vbox  entirely I simply commented out the code for the Vbox in HomePage as I didn't want to get rid of it without approval.